% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colordiverger.R
\name{colordiverger}
\alias{colordiverger}
\title{Create diverging color palette}
\usage{
colordiverger(
  color1 = "mediumpurple",
  color2 = "tan",
  min.val = -5,
  max.val = 5
)
}
\arguments{
\item{color1}{a color of R's predefined color names which is available in 4 graduations with names as 
follows for the example of cyan: cyan1, cyan2, cyan3, cyan4; 
you may check availability online or within R with 'colors()';
color1 defines the color for all negative values;
default: "mediumpurple"}

\item{color2}{a color with the same demands as defined for color1;
color2 defines the color for all positive values;
default: "tan"}

\item{min.val}{provide the minimum value of the dataset the palette will be prepared for;
if you set min.val to -5 and max.val to 5, the palette will follow z-score graduations;
default: -5}

\item{max.val}{provide the maximum value of the dataset the palette will be prepared for;
default: 5}
}
\value{
customized diverging color palette
}
\description{
A function that takes two colors as its input and produces a color palette for plotting 
(z-scored) values (e.g. in a heatmap) with two diverging colors.
}
\examples{
my_color_palette <- colordiverger("springgreen", "steelblue")
}
\author{
Paul Volkmann
}
