% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getexpdata.R
\name{getexpdata}
\alias{getexpdata}
\title{Get data of NTB experiments}
\usage{
getexpdata(
  directory,
  analysis = c("4arm", "2arm_tg", "2arm_ko", "2arm_sd"),
  ordercolumns = c("ntb", "rdoc", "manual"),
  ordercolumns_manual,
  exclude = FALSE
)
}
\arguments{
\item{'directory':}{file directory of Behavior and Animal List files
(mind correct spelling of both files and directory!)}

\item{'analysis':}{specifying the kind of experiment performed - 4-arm or 2-arm
with either transgenic or knock-out animals or with social defeat looling at environmental condition
as group of interest
(respectively, choosing the kind of analysis preferred)
(default: "4arm")}
}
\value{
prepared and joined dataframe of all animals and corresponding NTB experiments
}
\description{
A function that imports an NTB dataset and prepares the data for plotting and analysis.
For right formatting of your files, please consider the "ReadMe for ntbgraphics".
}
\author{
Paul Volkmann
}
