% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getexpdata_2arm.R
\name{getexpdata_2arm}
\alias{getexpdata_2arm}
\title{Get data of NTB experiments}
\usage{
getexpdata_2arm(directory)
}
\arguments{
\item{'directory':}{file directory of Behavior and Animal List files
(mind correct spelling of both files and directory!)}
}
\value{
prepared and joined dataframe of all animals and corresponding NTB experiments
}
\description{
A function that imports a 2-arm GT NTB dataset and prepares the data for plotting and analysis.
}
\author{
Paul Volkmann
}
