% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapexp.R
\name{heatmapexp}
\alias{heatmapexp}
\title{Heatmapping of NTB experiment dataset}
\usage{
heatmapexp(datamatrix, directory, title)
}
\arguments{
\item{'datamatrix':}{matrix of animals and corresponding NTB experiments (e.g. output of 'prepareheatmap')}

\item{'directory':}{file directory of Animal List file}

\item{'title':}{define the title of your heatmap}
}
\value{
heatmap
}
\description{
A function that produces a heatmap for an NTB dataset. (Warning: This function relies on function 'getexpdata'!)
}
\author{
Paul Volkmann
}
