% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapexp_2arm.R
\name{heatmapexp_2arm}
\alias{heatmapexp_2arm}
\title{Heatmapping of NTB experiment dataset (2-arm)}
\usage{
heatmapexp_2arm(datamatrix, directory, title)
}
\arguments{
\item{'datamatrix':}{matrix of animals and corresponding NTB experiments (e.g. output of 'prepareheatmap')}

\item{'directory':}{file directory of Animal List file}

\item{'title':}{define the title of your heatmap}
}
\value{
heatmap
}
\description{
A function that produces a heatmap for an NTB 2-arm dataset. (Warning: This function relies on function 'getexpdata_2arm'!)
}
\author{
Paul Volkmann
}
