% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopplotexp.R
\name{loopplotexp}
\alias{loopplotexp}
\title{Plot NTB experiments}
\usage{
loopplotexp(
  directory,
  analysis = c("4arm", "2arm_tg", "2arm_ko"),
  orderplots = c("other", "tcf4"),
  saveplotdir = directory
)
}
\arguments{
\item{'directory':}{file directory of Behavior and Animal List files}

\item{'analysis':}{specifying the kind of experiment performed - 4-arm by default or 2-arm
with either transgenic or knock-out animals as group of interest
(or choosing the kind of analysis preferred)}

\item{'orderplots':}{gives user ability to specifiy order of plots
(by default and for "other", plot order will depend on alphabetical order of GT_Env objects;
for "tcf4", plot order will be wt_hc, wt_sd, tg_hc, tg_sd (only for 4-arm experiments))}

\item{'saveplotdir':}{file directory where to save plots
(default at location of Behavior and Animal List files)}
}
\value{
boxplot saved as PDF
}
\description{
A function that takes an NTB dataset
and visualizes the data as customized boxplots.
(Requires functions 'ploteachexp' and 'getexpdata' internally.)
}
\author{
Paul Volkmann
}
