% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcatsneexp.R
\name{pcatsneexp}
\alias{pcatsneexp}
\title{Perform and plot PCA and tSNE of NTB experiment dataset}
\usage{
pcatsneexp(
  directory,
  analysis = c("4arm", "2arm_tg", "2arm_ko", "2arm_sd"),
  perplex,
  theta = 0.5,
  pastetitle = "PCA",
  pastetitle2 = "tSNE"
)
}
\arguments{
\item{'directory':}{file directory of Behavior and Animal List files}

\item{'analysis':}{specifying the kind of experiment performed - 4-arm or 2-arm
with either transgenic or knock-out animals or with social defeat looling at environmental condition
as group of interest
(respectively, choosing the kind of analysis preferred)
(default: "4arm")}

\item{'perplex':}{define perplexity parameter (should not be bigger than >[nrow(matrix] - 1)/3<)}

\item{'theta':}{define theta in the range of 0 to 1 (speed/accuracy trade-off;
increase for less accuracy, set to 0 for exact TSNE)
(default: 0.5)}

\item{'pastetitle':}{customize title of PCA plots; title for PCA will be "pastetitle",
title for PCA with ellipse will be "pastetitle Ellipse"
(default: "PCA")}

\item{'pastetitle2':}{define title of tSNE plot
(default: "tSNE")}
}
\value{
three cluster plots and a list containing the results
}
\description{
A function that produces cluster plots for results of Principal Component Analysis and
t-distributed Stochastic Neighbor Embedding of a z-scored NTB dataset
and returns a list containing the results (see examples for easy way to access individually).
(Requires function 'getexpdata' internally.)
For right formatting of your files, please consider the "ReadMe for ntbgraphics".
}
\author{
Paul Volkmann
}
