% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploteachexp_2arm.R
\name{ploteachexp_2arm}
\alias{ploteachexp_2arm}
\title{Plot NTB experiments (2-arm)}
\usage{
ploteachexp_2arm(expname, directory)
}
\arguments{
\item{'expname':}{name of a column/experiment of an NTB dataset}

\item{'directory':}{file directory where to save plots}
}
\value{
boxplot saved as PDF
}
\description{
A function that takes experiments (columns) of a 2-arm NTB dataset
and visualizes the data as customized boxplots.
}
\author{
Paul Volkmann
}
