% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareheatmap.R
\name{prepareheatmap}
\alias{prepareheatmap}
\title{Prepare data of NTB experiments for Heatmapping}
\usage{
prepareheatmap(dataframe)
}
\arguments{
\item{'dataframe':}{dataframe of animals and corresponding NTB experiments (e.g. output of 'getexpdata')}
}
\value{
prepared matrix of z-scored behavioral measures
}
\description{
A function that prepares NTB data for heatmap analysis including z-scoring.
}
\author{
Paul Volkmann
}
