% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getexpdata.R
\name{getexpdata}
\alias{getexpdata}
\title{Get data of NTB experiments}
\usage{
getexpdata(
  directory,
  analysis = c("2arm_ko", "2arm_tg", "2arm_sd", "2arm_treat", "4arm_sd_ko",
    "4arm_sd_tg", "4arm_treat_ko", "4arm_treat_tg"),
  ordercolumns = c("ntb", "rdoc", "manual"),
  ordercolumns_manual,
  exclude.animals = FALSE,
  orderlevelcond = c("other", "gtblock", "etblock")
)
}
\arguments{
\item{directory}{file directory of 'Meta Behavior' and 'Animal List' files
(mind correct spelling of both files and directory!)
(no default)}

\item{analysis}{specifying the kind of experiment performed;
"4arm_sd_tg", "4arm_sd_ko", "4arm_treat_tg", "4arm_treat_ko",
"2arm_tg", "2arm_ko", "2arm_sd", "2arm_treat";
(tg for transgenic, ko for knockout;
4arm_sd_x assumes a stress paradigm with social defeat (sd) and housing or handling control (hc) as 
control;
4arm_treat_x assumes a treatment paradigm with treated (treat) and untreated (untreat) animals;
2arm_x assumes wildtype controls (wt) for tg and ko, housing or handling controls (hc) for sd and
untreated controls (untreat) for treat
(analysis defines the kind of experiment performed, respectively the kind of analysis preferred - 
you can easily perform 2arm analysis for 4arm experiments, looking only at the groups of interest, 
but not the other way around)
(default: "2arm_ko")}

\item{ordercolumns}{defining the order of experiment column appearance in final table;
RFID and Condition are always listed first and need no specification;
order of experiments may be chronological with "ntb", follow RDoC clustering with "rdoc" or be customized
manually with "manual" (-> use ordercolumns_manual for exact appearance; there, you may also choose to 
exclude experiments)
(default: "ntb")}

\item{ordercolumns_manual}{customizing order of appearance and appearance itself of experiment columns 
in final table (experiments that are not listed will not be included);
only if ordercolumns = "manual";
user has to provide a vector containing characters (e.g. by using c("Meanspeed", "SerialLearn")) 
with all experiments he wants to include into the final tabel
(no default; no need for specification if ordercolumns is not "manual")}

\item{exclude.animals}{excluding animals from analysis by RFID
user has to provide a vector containing characters (e.g. by using c("900200000067229", "900200000065167")) 
with all animals he wants to exclude from the final table;
if FALSE is provided, no animal will be excluded
(default= FALSE)}

\item{orderlevelcond}{defines order of factor levels of conditions 
(which might be important when it comes to plotting or displaying your data grouped by condition 
in a defined order):
"other" for alphabetical order in case of 4arm and >'control', 'condition'< as default order for 
2arm experiments;
"gtblock" for order wt_x, wt_y, tg_x, tg_y;
"etblock" for order x_hc, y_hc, x_sd, y_sd;
note that no (further) specification is possible for 2arm experiments
(default: "other")}
}
\value{
prepared and joined dataframe of all animals and corresponding NTB experiments
}
\description{
A function that imports an NTB dataset and prepares the data for plotting and analysis.
For right formatting of your files, please consider the "ReadMe for ntbgraphics".
}
\author{
Paul Volkmann
}
