% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getexpdata.R
\name{getexpdata}
\alias{getexpdata}
\title{Get data of NTB experiments}
\usage{
getexpdata(directory)
}
\arguments{
\item{'directory':}{file directory of Behavior and Animal List files
(mind correct spelling of both files and directory!)}
}
\value{
prepared and joined dataframe of all animals and corresponding NTB experiments
}
\description{
A function that imports a 4-arm GT_Env NTB1.0 dataset and prepares the data for plotting and analysis.
}
\author{
Paul Volkmann
}
