% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapexp.R
\name{heatmapexp}
\alias{heatmapexp}
\title{Heatmapping of NTB experiment dataset}
\usage{
heatmapexp(
  directory,
  analysis = c("4arm", "2arm_tg", "2arm_ko"),
  title = "Heatmap"
)
}
\arguments{
\item{'directory':}{file directory of Behavior and Animal List files}

\item{'analysis':}{specifying the kind of experiment performed - 4-arm by default or 2-arm
with either transgenic or knock-out animals as group of interest
(or choosing the kind of analysis preferred)}

\item{'title':}{define the title of your heatmap (default: "Heatmap")}
}
\value{
heatmap and z-scored data matrix
}
\description{
A function that produces a heatmap for an NTB dataset and a data matrix with z-scored values.
(Requires function 'getexpdata'internally.)
}
\author{
Paul Volkmann
}
