% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopplotexp.R
\name{loopplotexp}
\alias{loopplotexp}
\title{Plot every NTB experiment into one single PDF}
\usage{
loopplotexp(
  directory,
  analysis = c("4arm", "2arm_tg", "2arm_ko", "2arm_sd"),
  orderplots = c("other", "tcf4"),
  saveplotdir = directory
)
}
\arguments{
\item{'directory':}{file directory of Behavior and Animal List files}

\item{'analysis':}{specifying the kind of experiment performed - 4-arm or 2-arm
with either transgenic or knock-out animals or with social defeat looling at environmental condition
as group of interest
(respectively, choosing the kind of analysis preferred)
(default: "4arm")}

\item{'orderplots':}{gives user ability to specifiy order of plots
(for "other", plot order will depend on alphabetical order of GT_Env objects;
for "tcf4", plot order will be wt_hc, wt_sd, tg_hc, tg_sd (only for 4-arm experiments))
(default: "other")}

\item{'saveplotdir':}{file directory where to save plots
(default: location of Behavior and Animal List files as specified in 'directory')}
}
\value{
all boxplots saved in PDF
}
\description{
A function that takes an NTB dataset
and visualizes the data as customized boxplots.
(Requires functions 'ploteachexp' and 'getexpdata' internally.)
For right formatting of your files, please consider the "ReadMe for ntbgraphics".
}
\author{
Paul Volkmann
}
