% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploteachexp.R
\name{ploteachexp}
\alias{ploteachexp}
\title{Plot NTB experiments}
\usage{
ploteachexp(expname, directory)
}
\arguments{
\item{'expname':}{name of a column/experiment of an NTB dataset}

\item{'directory':}{file directory where to save plots}
}
\value{
boxplot saved as PDF
}
\description{
A function that takes experiments (columns) of a 4-arm GT_Env NTB dataset
and visualizes the data as customized boxplots.
}
\author{
Paul Volkmann
}
